Close curly brace, improve readability in README
There was a closing curly brace missing on the example `pip install`
line, so this change adds that brace.  But also, the undifferentiated
string "yourgithubforkurl" had the potential to fool the eye -- when I
was just skimming, I at first read it as "your github for kurl", and
thought the last word was an intentional fun misspelling of "curl".
Then when I looked more closely I realized what the string meant.  So
this change also puts the string in StudlyCaps to improve readability.