Add Leaflet Map Component
This commit installs the leafletjs[1] package to the repository, which
we have decided to use(for now) to serve the basic rendering of
artwork's as points on a map, in conjunction with OpenStreetMap. It adds
the LeafletMap component which is then rendered side-by-side the
existing list of resources on the home page. Currently, the component
only renders a map with tiles from open street map, following the setup
instructions on the leaflet home page.

[1] https://leafletjs.com/index.html