Add Pinia Store
This commit introduces Pinia to the project, the officially
reccommended[1] state management plugin for Vue.js. The thought behind
Pinia's inclusion is to use it to store a global 'resourceId' reference,
which can be used to change the content of a resource viewer panel(to be
added in subsequent commits). Longter,, as the project progresses, it
will also likely be useful to have access to a global state for certain
key data points.

[1] https://pinia.vuejs.org/