Add Resource Panel Interactivity
This commit enables users to get details about artworks or artists in
the resource panel, by clicking either a map icon or the 'show details'
button in the search list. The resource panel is ultimately intended to
show both resource data and link to related resources, but resource
relations will need to be further defined as types and validated, so
that will be handled in a subsequent PR. Furthermore, this commit also
builds the latest front-end code and places it in the django
application's static directory. Long-term, a more thorough and seperate
build process should be implemented.