Add Resource type, refactor other types
This commit refactors the existing data types to more accurately
reflect the data recieved via the arches 'resources/{resourceid}'
endpoint. Principally, it abstracts what is common between both
Artist and Artwork data and defines that in the new Resource type,
and clarifies the existing Artist and Artwork types to be only their
unique fields. This will also enable a more fluid system of displaying
the metadata of each resource in the subsequent commit, by allowing the
resource panel to display meta-data agnostic of resource type