Add Custom Django Artwork Map Route
This commit adds two new custom routes to the arches application, one
that provides the existing graphs in the project, and one that returns
a list of all artwork data combined with it's geojson data. This was
decided on after concluding that the existing REST API infrastructure
provided in base arches could not serve the goal of gathering all
Artworks in a single API call. It leverages the Django ORM and shared
application resource store, and is notably the first time the project
has directly pulled data from the arches application, as opposed to just
consuming it on the front-end.