Add Coordinates Type
This commit adds the Coordinates type to the project, which is made to
represent the parsed json from the coordinates field of an Artwork
Resource from Arches. Currently, in the WAC arches instance, the
coordinates field is returned from the arches API as a string literal,
so for the time being this serves as a way to use typescript on location
data. This will be useful for the map and detail components moving
forward