Add All Artworks to Map
This commit modifies the Leaflet Map component to consume
data from the new 'archesdataviewer/artworks' route, allowing
for all artworks to be loaded into the map on page load. Since
the route provides the coordinate data as a string rather than
a json object, it also adds a utility function to parse this string
into json. Longterm this method should be replaced once we can ensure
that the data being passed from the backend is proper json.