Add image route
This commit adds a new django route, 'getImage', which returns a list of
image data for either an artist or artwork resource. In order to display
an image for an Artist, which does not have an image related to it, we
need to fetch from the related Artworks first, and from there display an
artwork image.  Artworks are directly related to images, so theirs can
be yielded from a direct Tile lookup by nodeid. The route decisions the
resource type against the graphId, returning a list of tile data as a
JSONResponse, or an empty list if no resource matching the passed
resource_id can be found