Various Demo Tidyings
This commit contains a series of small changes to the repository that
ere done for an application demo, but are worth keeping in the master
repository. All changes were user-facing, which is to say they only
served to tidy up the appearance of the application. The changes were:

1) Not showing rose mcgowan as an artist, which is done to avoid the
Issue where she has a relation to all artworks.
2) Add Empty Field Handling, which displays a default value when there is no data present.
3) Fit Content To Screen
4) Credit to @kfogel:
   "Use a self-explanatory "No image available." image
   Instead of showing a "?"  as the thumbnail when an artwork has no
   associated image, show gray text saying "No image available."
   Coincidentally, this wording matches the alt text that @hminsky2002
   already had in 'front-end/src/components/ResourceListItem.vue' for
   this image, although I did not know that at the time I wrote it.

   Along with the PNG, I include the Gimp XCF source here, to make it
   easy for us to change this wording if we ever want to."

5) Add modal image view, which displays an artwork at full size on click
of its image
6) Check-in latest front-end assets
7) Change For Deploy, which adds some hard public assets to the
django application folder to get around nginx issues