Use a self-explanatory "No image available." image
Instead of showing a "?" as the thumbnail when an artwork has no
associated image, show gray text saying "No image available."
Coincidentally, this wording matches the alt text that @hminsky2002
already had in 'front-end/src/components/ResourceListItem.vue' for
this image, although I did not know that at the time I wrote it.

Along with the PNG, I include the Gimp XCF source here, to make it
easy for us to change this wording if we ever want to.