Accommodate WAC canonical data changes
This commit updates the repository to accommodate the latest changes to
the canonical wac dataset, now hosted here
https://code.librehq.com/ots/arches/arches-demo-data. These changes are
principally reflected in the refactor of the Artwork resource type, and
the getImage route (as the canonical node ids for an artworks image url
file has changed). Thankfully, since the Resource typescript types were
designed to contain only data shared across all arches resources, we do
not need to do any more refactoring than the datatypes that we ourselves
define.

One non-breaking change is that since the canonical dataset does not
currently provide geojson data for Artworks, the map component in the
vue application does not render any artworks. However, as soon as we do
add geojson back to the artworks, it will render.