Add Basic Search Functionality
This commit adds a rudimentary search capability to the searchlist
component, which allows for resource retrieval by name or resource model
type. This is done via calls to the elasticsearch node using the same
`/search/resources` route already being used by the component. A new
search is made on each keystroke which could have scaling issues in the
future, but elastic search seems to be able to maintain [1]. As arches
is serving as a connector to the elasticsearch node, it will do good to
further conform to elasticsearch design patterns as we develop search on
the frontend

[1]https://www.elastic.co/guide/en/elasticsearch/reference/current/search-as-you-type.html