Remove all styling and stylistic components
This commit removes most css from the project, as we are moving to an
entirely new design setup, and these styles were only meant to be
temporary. It retains only the css required to render each panel
in the home page, which will also be modified in subsequent commits.
Finally, it gets rid of the modal image component, and next and previous
search page components, as these will not be used in the final design.