Refactor data fetching pattern
This commit totally refactors our data fetching pattern for the
application, adopting a prefetch approach to all resources instead of
relying on the search endpoint in arches. Since we are working with a
relatively small dataset, we are able to take advantage of preloading
all resources on the page without having to worry about repetead api
calls to arches. Furthermore, with the introduction of the prefetched
reactive variables added to our home component, we're able to use this
prefetched data not just to load resources, but to also handle matching
resources with image tiles, and coordinate data, negating the need for
our custom getImage and artworkMap routes.

Furthermore, this commit drops the resource relation components and
multiple provider components, as the only subsequent api calls made by
the application after prefetch happen in the resource detail
component. This is due to the fact that we are only retrieving a
reference to the resources in our prefetch, rather than the deeply
populated resource themselves. Retrieving references versus the full
resources improved API response time by a factor of 10, and seems to be
a preferable approach for fetching data.