Add Resource Rendering for all resource types
This commit adds rendering for images and details for resources to the
ResourcePanel, which is brought up on click of a resource in the search
list. In order to accommodate rendering artwork images on non-artwork
pages, the new ResourceDetail and ResourceDetailItem components have
been created using the vue slot tempting syntax.  Furthermore, to
actually retrieve the data for non-artwork resources, we reintroduce
usage of the 'resource/related' route in base arches, which we now fetch
from for both the ResourcePanelProvider and RelatedArtworkDetail
component.