Add more extensive linting packages
This commit adds the official vue/eslint-airbnb-typescript package to
the project, along with the eslint-plugin-vue-scoped-css package, to
further extend the linting process for the application. Notably, since
the eslint-airbnb-with-typescript package does not support the
flat-config format for an eslint config file, we have switched to an
eslintrc.cjs file format.