Introduce vue router
This commit introduces the vue router package to the application, which
is the official client-side routing handler for vue. We do so to add
routing to the application, which is used to navigate between resources
based on resource id. Resource id is now handled as a routing parameter,
instead of through pinia, which offers a much more fluid rendering
process. As such, pinia has been removed from the application until such
time as it is needed again.