Remove structure and photographer filters
This commit removes the structure and photographer filters in the search
list, as well as defaults the resource to display artworks.  This is
done as part of the overall effort to not show structure or photographer
resource models, for the time being, to streamline the user experience
and emphasize just artwork and artists in the wac dataset. We may
reintroduce these as part of the UI in the future, but for now they are
not part of the demo.