Fix resource update from map icons
This commit fixes an issue where resources in the resource panel were
not being updated when fetched from the map. This was due to a lack
of reactivity in the component props, which has been mitigated using
vue's computed property, which allows for preserving reactivity
on resources within other reactive resources. neat!