Add map flyover and highlight for active resource
This commit adds flyover to the map to the active resource based on the
route id parameter. It also pulls out the map configuration into some
variables to make customization of default zoom levels simpler, along
with refactoring the icon image path to support local and production
environments, along with trimming out unecessary old static icon
images. Furthermore, it adds an active icon and functionality to switch
from the default to the active icon on new active resource id